requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
sqlmodel==0.0.24
SQLAlchemy==2.0.41
python-dotenv==1.1.0
pydantic==2.11.5
python-multipart==0.0.20
pyjwt==2.10.1
passlib==1.7.4
bcrypt==4.3.0
psycopg2-binary==2.9.10
email_validator==2.2.0
pytest==8.2.2
httpx==0.27.0